+++ src/experimental/lib/effect-manager.js
