+++ src/lib/index.js
@@ -5,8 +5,6 @@ import './init';
 export {default as Layer} from './layer';
 export {default as AttributeManager} from './attribute-manager';
 export {default as LayerManager} from './layer-manager';
-export {default as EffectManager} from './effect-manager';
-export {default as Effect} from './effect';
 
 // Object iteration helper
 export * from './utils/object-iterator';
